[refactor] Entity패키지로 이동 및 어노테이션 추가
- NotificationEntity를 Entity 패키지로 이동
- Entity ID에 nullable=false 추가
